requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
ccxt==4.0.0
yfinance==0.2.32
ta==0.10.2
requests==2.31.0
celery==5.3.4
redis==5.0.1
aiohttp==3.9.1
httpx==0.25.1
python-multipart==0.0.6
